refactor: remove duplicate and dead modules, bump version to 3.6.0
- Delete qxub/shortcut_manager.py (exact duplicate of config/shortcuts.py)
  Update 11 imports across exec_cli, config_cli, interactive_cli
- Delete qxub/platform_cli.py (dead code, superseded by platforms/cli.py)
- Delete qxub/execution_context.py (legacy, superseded by execution/context.py)
  Update import in remote/command_builder.py
- Remove broken imports from remote/__init__.py referencing non-existent
  modules (.config, .core, .executor, .loader)
- Bump version to 3.6.0 in __init__.py and setup.py (fixes prior mismatch)
- Net: -1051 lines of dead/duplicate code
